+++ backend/src/controllers/ValidateTokenController.ts
@@ -1,12 +1,12 @@
 import { Request, Response } from 'express';
 import { TokenHelper } from '../helpers/TokenHelper.js';
-import { log } from '../logger.js';
 import { datasource } from '../helpers/DatabaseHelper.js';
 import { UserNotFoundError } from '../errors/UserNotFoundError.js';
 import { User } from '../entities/User.js';
 import { Team } from '../entities/Team.js';
 import { TeamNotFoundError } from '../errors/TeamNotFoundError.js';
 import { ObjectId } from 'mongodb';
+import { log } from '../worker.js';
 
 const validate = async (req: Request, res: Response) => {
   log.debug(`validate token ${req.body.token}`);
@@ -14,19 +14,13 @@ const validate = async (req: Request, res: Response) => {
   try {
     const payload = TokenHelper.verifyJwt(req.body.token);
 
-    const team = await datasource.manager.findOneById(
-      Team,
-      new ObjectId(payload.teamId)
-    );
+    const team = await datasource.manager.findOneById(Team, new ObjectId(payload.teamId));
 
     if (!team) {
       throw new TeamNotFoundError();
     }
 
-    const user = await datasource.manager.findOneById(
-      User,
-      new ObjectId(payload.userId)
-    );
+    const user = await datasource.manager.findOneById(User, new ObjectId(payload.userId));
 
     if (!user) {
       throw new UserNotFoundError();